+++ example/lib/widgets/skip_file_1.dart
@@ -1,0 +1,6 @@
+class SkipFromThisClass {
+  static final appID = "com.example.code";
+  static final appSecret = "TWQEW-5FASR-FAGSE-532FSG";
+  static final apiToken =
+      "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsImlhdCI6MTUxNjIzOTAyMn0.KMUFsIDTnFmyG3nMiGM6H9FNFUROf3wh7SmqJp-QV30";
+}